+++ src/Codenizer.HttpClient.Testable/ResponseConfigurationException.cs
@@ -1,0 +1,15 @@
+using System;
+
+namespace Codenizer.HttpClient.Testable
+{
+    /// <summary>
+    /// Thrown when the <see cref="RequestBuilder"/> has not been fully configured
+    /// </summary>
+    public class ResponseConfigurationException : Exception
+    {
+        internal ResponseConfigurationException(string message) : base(message)
+        {
+            
+        }
+    }
+}