+++ test/Codenizer.HttpClient.Testable.Tests.Unit/WhenConfiguringHandler.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Net.Http;
 using Xunit;
 using FluentAssertions;
 
@@ -35,7 +34,8 @@ namespace Codenizer.HttpClient.Testable.Tests.Unit
                 client.GetAsync("https://tempuri.org/api/hello").GetAwaiter().GetResult();
             }
             catch
-            {   
+            {
+                // ignored
             }
 
             handler
@@ -50,8 +50,8 @@ namespace Codenizer.HttpClient.Testable.Tests.Unit
             var handler = new TestableMessageHandler();
             var client = new System.Net.Http.HttpClient(handler);
 
-            handler.RespondTo("/api/hello");
-            handler.RespondTo("/api/hello");
+            handler.RespondTo().Get().ForUrl("/api/hello");
+            handler.RespondTo().Get().ForUrl("/api/hello");
 
             Action action = () => client.GetAsync("https://tempuri.org/api/hello").GetAwaiter().GetResult();
 
@@ -64,8 +64,8 @@ namespace Codenizer.HttpClient.Testable.Tests.Unit
             var handler = new TestableMessageHandler();
             var client = new System.Net.Http.HttpClient(handler);
 
-            handler.RespondTo("/api/hello?foo=bar");
-            handler.RespondTo("/api/hello?foo=bar");
+            handler.RespondTo().Get().ForUrl("/api/hello?foo=bar");
+            handler.RespondTo().Get().ForUrl("/api/hello?foo=bar");
 
             Action action = () => client.GetAsync("https://tempuri.org/api/hello?foo=bar").GetAwaiter().GetResult();
 
@@ -78,8 +78,8 @@ namespace Codenizer.HttpClient.Testable.Tests.Unit
             var handler = new TestableMessageHandler();
             var client = new System.Net.Http.HttpClient(handler);
 
-            handler.RespondTo("/api/hello?foo=bar");
-            handler.RespondTo("/api/hello?foo=qux");
+            handler.RespondTo().Get().ForUrl("/api/hello?foo=bar");
+            handler.RespondTo().Get().ForUrl("/api/hello?foo=qux");
 
             Action action = () => client.GetAsync("https://tempuri.org/api/hello?foo=bar").GetAwaiter().GetResult();
 
@@ -92,8 +92,8 @@ namespace Codenizer.HttpClient.Testable.Tests.Unit
             var handler = new TestableMessageHandler();
             var client = new System.Net.Http.HttpClient(handler);
 
-            handler.RespondTo(HttpMethod.Get, "/api/hello?foo=bar");
-            handler.RespondTo(HttpMethod.Get, "/api/hello?foo=bar");
+            handler.RespondTo().Get().ForUrl("/api/hello?foo=bar");
+            handler.RespondTo().Get().ForUrl("/api/hello?foo=bar");
 
             Action action = () => client.GetAsync("https://tempuri.org/api/hello?foo=bar").GetAwaiter().GetResult();
 
@@ -106,12 +106,44 @@ namespace Codenizer.HttpClient.Testable.Tests.Unit
             var handler = new TestableMessageHandler();
             var client = new System.Net.Http.HttpClient(handler);
 
-            handler.RespondTo(HttpMethod.Get, "/api/hello?foo=bar");
-            handler.RespondTo(HttpMethod.Post, "/api/hello?foo=bar");
+            handler.RespondTo().Get().ForUrl("/api/hello?foo=bar");
+            handler.RespondTo().Post().ForUrl("/api/hello?foo=bar");
 
             Action action = () => client.GetAsync("https://tempuri.org/api/hello?foo=bar").GetAwaiter().GetResult();
 
             action.Should().NotThrow<MultipleResponsesConfiguredException>();
+        }
+
+        [Fact]
+        public void GivenRequestBuilderWithMethodButNoUrl_ResponseConfigurationExceptionIsThrown()
+        {
+            var handler = new TestableMessageHandler();
+            handler.RespondTo().Get();
+            var client = new System.Net.Http.HttpClient(handler);
+
+            Action action = () => client.GetAsync("https://tempuri.org/api/hello?foo=bar").GetAwaiter().GetResult();
+
+            action.Should().Throw<ResponseConfigurationException>()
+                .Which
+                .Message
+                .Should()
+                .Be("The URL to respond to has not been set");
+        }
+
+        [Fact]
+        public void GivenRequestBuilderWithout_ResponseConfigurationExceptionIsThrown()
+        {
+            var handler = new TestableMessageHandler();
+            handler.RespondTo().ForUrl("/derp");
+            var client = new System.Net.Http.HttpClient(handler);
+
+            Action action = () => client.GetAsync("https://tempuri.org/api/hello?foo=bar").GetAwaiter().GetResult();
+
+            action.Should().Throw<ResponseConfigurationException>()
+                .Which
+                .Message
+                .Should()
+                .Be("The HTTP verb to respond to has not been set");
         }
     }
 }