+++ Changelog.md
@@ -1,5 +1,38 @@
 # Codenizer.HttpClient.Testable Changelog
 
+## 2.1.0
+
+This release introduces a more fluent way to configure the responses. Instead of calling the handler like so:
+
+```csharp
+handler
+ .RespondTo(HttpMethod.Post, "/api/some/endpoint")
+ .With(HttpStatusCode.Created);
+```
+
+you can now write this as:
+
+```csharp
+handler
+ .RespondTo()
+ .Post()
+ .ForUrl("/api/some/endpoint")
+ .With(HttpStatusCode.Created);
+```
+
+The value of the `Content-Type` header can be set as well using `AndContentType`:
+
+```csharp
+handler
+ .RespondTo()
+ .Post()
+ .ForUrl("/api/some/endpoint")
+ .AndContentType("application/json")
+ .With(HttpStatusCode.Created);
+```
+
+The `RespondTo(HttpMethod method, string url, string contentType)` method is marked as obsolete but as a warning. It will be removed in version 3.x in the future.
+
 ## 2.0.0
 
 - Added XML Doc comments on public methods and properties to improve documentation in the IDE